refactor: extract form HTML to views file
